Add a reproducer for p11-kit bug when adding new devices
This program loads a key pair using the engine pkcs11, forks to create
a new process, and waits for a SIGUSR1 signal before trying to
sign/verify random data in both parent and child processes.

The intention of the signal waiting is to allow the user to add/remove
devices before continuing to the signature/verifying test.

Signed-off-by: Anderson Toshiyuki Sasaki <ansasaki@redhat.com>